+++ src/main.ts
@@ -1,7 +1,7 @@
 import { createApp } from "vue";
 import { createStore } from "./store";
 import { router } from "./router";
-import App from "./layouts/default.vue";
+import App from "./App.vue";
 import "./assets/css/base.css";
 
 createApp(App).use(createStore()).use(router).mount("#app");
